ad7172: Correct postfilter rates and rejection
The AD7172-2 Rev. A datasheet is inconsistent on the 50/60 Hz postfilter
rates and rejection figures between the *ENHANCED 50 Hz AND 60 Hz
REJECTION FILTERS* and the *FILTER CONFIGURATION REGISTER 0* sections.
Use the former section's figures instead.

The latter section is wrong since the 21.25 SPS figure is incorrect, as
the output data rate = 1 / (settling time), and 1 / (40 ms) = 25 Hz.
This seems to have been corrected in the datasheets for the rest of the
AD717x family, which have the same filter.